+++ .scripts/start-autorest-wiremock.js
@@ -46,6 +46,7 @@ async function main() {
   const command = cmdArgs.shift()
 
   let serverProc = undefined;
+  let cmdProc = undefined;
   let spResolve = undefined;
   let spReject = undefined;
   let running = false;
@@ -69,8 +70,16 @@ async function main() {
   // start the wiremock process
   verbose('Starting WireMock.')
   const spResult = execute(process.execPath, [`${wmFolderPath}/node_modules/wiremock/jdeploy-bundle/jdeploy.js`, '--root-dir', resolve(`${__dirname}/..`), '--port', '3000', '--verbose'], {
-    onCreate: (process) => {
-      serverProc = process;
+    onCreate: (proc) => {
+      serverProc = proc;
+      proc.on('close', ()=>{
+        if( cmdProc && cmdProc.status === null ) {
+          cmdProc.kill();
+        }
+        if( interactive ) {
+          process.exit(0);
+        }
+      })
     },
     onStdOutData: (chunk) => {
       const c = chunk.toString().replace(/\s*$/, '');
@@ -97,6 +106,7 @@ async function main() {
     await execute(command, cmdArgs, {
       onCreate: (process) => {
         verbose(`Started command: '${command} ${cmdArgs.join(' ')}'`);
+        cmdProc = process;
         running = true;
       },
       onStdOutData: (chunk) => {
@@ -113,6 +123,7 @@ async function main() {
   } else {
     await queryUser('\n\nPress enter to stop testserver\n');
   }
+  cmdProc = undefined;
 
   // after the cmdline is done.
   // shutdown server process
@@ -123,7 +134,7 @@ async function main() {
   verbose('Waiting for WireMock to finish.');
 
   // force-kill server process
-  if (serverProc.status === null) {
+  if (serverProc && serverProc.status === null) {
     verbose('killing WireMock');
     serverProc.kill();
   }
